// packages/core/src/schema/types/config/access-control.ts
export type BaseAccessArgs<ListTypeInfo extends BaseListTypeInfo> = {
  session: any;
  listKey: string;
  context: PickerContextFromListTypeInfo<ListTypeInfo>;
};

export type AccessOperation = 'create' | 'query' | 'update' | 'delete';
export type FilterOperation = 'query' | 'update' | 'delete';
export type ItemOperation = 'create' | 'update' | 'delete';

export type ListOperationAccessControl<Operation extends AccessOperation, ListTypeInfo extends BaseListTypeInfo> = (
  args: BaseAccessArgs<ListTypeInfo> & { operation: Operation }
) => MaybePromise<boolean>;

export type ListFilterAccessControl<Operation extends FilterOperation, ListTypeInfo extends BaseListTypeInfo> = (
  args: BaseAccessArgs<ListTypeInfo> & { operation: Operation }
) => MaybePromise<boolean | ListTypeInfo['inputs']['where']>;

export type CreateListItemAccessControl<ListTypeInfo extends BaseListTypeInfo> = (
  args: BaseAccessArgs<ListTypeInfo> & {
    operation: 'create';

    /**
     * The input passed in from the GraphQL API
     */
    inputData: ListTypeInfo['inputs']['create'];
  }
) => MaybePromise<boolean>;

export type UpdateListItemAccessControl<ListTypeInfo extends BaseListTypeInfo> = (
  args: BaseAccessArgs<ListTypeInfo> & {
    operation: 'update';

    /**
     * The item being updated
     */
    item: ListTypeInfo['item'];

    /**
     * The input passed in from the GraphQL API
     */
    inputData: ListTypeInfo['inputs']['update'];
  }
) => MaybePromise<boolean>;

export type DeleteListItemAccessControl<ListTypeInfo extends BaseListTypeInfo> = (
  args: BaseAccessArgs<ListTypeInfo> & {
    operation: 'delete';

    /**
     * The item being deleted
     */
    item: ListTypeInfo['item'];
  }
) => MaybePromise<boolean>;